requirements.txt
playwright
httpx
python-dotenv
beautifulsoup4